@@ -1,47 +1,28 @@
 {
-  "name": "client",
-  "version": "0.1.0",
+  "name": "dnaaaaa",
   "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@babel/core": "^7.10.4",
-    "@stripe/stripe-js": "^1.7.0",
-    "@vue/cli-plugin-babel": "~4.4.0",
-    "@vue/cli-plugin-eslint": "~4.4.0",
-    "@vue/cli-service": "~4.4.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "autoprefixer": "^9.8.5",
-    "axios": "^0.19.2",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
-    "core-js": "^3.6.5",
-    "css-loader": "^3.6.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^6.2.2",
-    "localforage": "^1.7.4",
-    "particles.js": "^2.0.0",
-    "prettier": "^1.19.1",
-    "qs": "^6.9.4",
-    "sass": "^1.26.10",
-    "sass-loader": "^9.0.2",
-    "style-loader": "^1.2.1",
-    "superagent": "^5.3.1",
-    "tailwindcss": "^1.4.6",
-    "vue": "^2.6.11",
-    "vue-embed-gist": "^1.2.0",
-    "vue-router": "^3.3.4",
-    "vue-svg-loader": "^0.16.0",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.6.2",
-    "vuex-persist": "^2.2.0"
+    "naive-ui": "^2.28.2",
+    "normalize.css": "^8.0.1",
+    "pinia": "^2.0.13",
+    "sass": "^1.50.1",
+    "vue": "^3.2.25",
+    "vue-router": "4"
   },
   "devDependencies": {
-    "style-resources-loader": "^1.3.2",
-    "vue-cli-plugin-style-resources-loader": "~0.1.4"
+    "@vitejs/plugin-vue": "^2.3.1",
+    "eslint": "^7.32.0 || ^8.2.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-vue": "^8.7.1",
+    "vite": "^2.9.5",
+    "vite-svg-loader": "^3.2.0"
   }
 }
